Revert "metodo EF para visualizarlo datos en la tabla"
This reverts commit 5df73129df2719e707e73ebc2e0a00c9f04a54bb.

+++ TiendaRopa/TiendaRopa/dtprenda.cs
@@ -19,6 +19,6 @@ namespace TiendaRopa
         public string marca { get; set; }
         public int talla { get; set; }
         public string color { get; set; }
-        public int precio { get; set; }
+        public decimal precio { get; set; }
     }
 }
